+++ src/runtime/components/Chip.vue
@@ -33,6 +33,12 @@ export interface ChipProps {
    */
   position?: Chip['variants']['position']
   /**
+   * If set to `true` the color is inverted.
+   * Used for 'air-primary', 'air-primary-success', 'air-primary-alert', 'air-primary-copilot' and 'air-primary-warning' colors.
+   * @defaultValue false
+   */
+  inverted?: boolean
+  /**
    * When `true`, keep the chip inside the component for rounded elements
    * @defaultValue false
    */
@@ -71,6 +77,7 @@ import { tv } from '../utils/tv'
 defineOptions({ inheritAttrs: false })
 
 const props = withDefaults(defineProps<ChipProps>(), {
+  inverted: false,
   inset: false,
   standalone: false,
   hideZero: false
@@ -86,6 +93,7 @@ const b24ui = computed(() => tv({ extend: tv(theme), ...(appConfig.b24ui?.chip |
   color: props.color,
   size: props.size, // size.value
   position: props.position,
+  inverted: Boolean(props.inverted),
   inset: Boolean(props.inset),
   standalone: Boolean(props.standalone),
   hideZero: Boolean(props.hideZero),