+++ CHANGELOG.md
@@ -93,10 +93,10 @@
   - fix color
   - add prop hideZero
   - add prop trailingIcon
+  - add prop inverted
   - add color
-    - air-primary, air-primary-inverted, air-primary-success, air-primary-success-inverted, air-primary-alert, air-primary-alert-inverted, air-primary-warning, air-primary-copilot, air-primary-no-accent, air-primary-no-accent-inverted
-    - air-secondary, air-secondary-alert, air-secondary-accent, air-secondary-accent-1, air-secondary-accent-2, air-secondary-no-accent
-    - air-tertiary, air-tertiary-accent, air-tertiary-no-accent
+    - air-primary, air-primary-success, air-primary-alert, air-primary-warning, air-primary-copilot,
+    - air-secondary, air-secondary-accent, air-secondary-accent-1, air-secondary-no-accent
   - @deprecate color
     - default, danger, success, warning, primary, secondary, collab, ai
   - @deprecate size